// src/lib/killbill/api/models/Tenant.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kill Bill
 * Kill Bill is an open-source billing and payments platform
 *
 * The version of the OpenAPI document: 0.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { AuditLog } from './AuditLog';
import { AuditLogFromJSON, AuditLogFromJSONTyped, AuditLogToJSON } from './AuditLog';

/**
 *
 * @export
 * @interface Tenant
 */
export interface Tenant {
	/**
	 *
	 * @type {string}
	 * @memberof Tenant
	 */
	tenantId?: string;
	/**
	 *
	 * @type {string}
	 * @memberof Tenant
	 */
	externalKey?: string;
	/**
	 *
	 * @type {string}
	 * @memberof Tenant
	 */
	apiKey: string;
	/**
	 *
	 * @type {string}
	 * @memberof Tenant
	 */
	apiSecret: string;
	/**
	 *
	 * @type {Array<AuditLog>}
	 * @memberof Tenant
	 */
	auditLogs?: Array<AuditLog>;
}

/**
 * Check if a given object implements the Tenant interface.
 */
export function instanceOfTenant(value: object): boolean {
	let isInstance = true;
	isInstance = isInstance && 'apiKey' in value;
	isInstance = isInstance && 'apiSecret' in value;

	return isInstance;
}

export function TenantFromJSON(json: any): Tenant {
	return TenantFromJSONTyped(json, false);
}

export function TenantFromJSONTyped(json: any, ignoreDiscriminator: boolean): Tenant {
	if (json === undefined || json === null) {
		return json;
	}
	return {
		tenantId: !exists(json, 'tenantId') ? undefined : json['tenantId'],
		externalKey: !exists(json, 'externalKey') ? undefined : json['externalKey'],
		apiKey: json['apiKey'],
		apiSecret: json['apiSecret'],
		auditLogs: !exists(json, 'auditLogs')
			? undefined
			: (json['auditLogs'] as Array<any>).map(AuditLogFromJSON)
	};
}

export function TenantToJSON(value?: Tenant | null): any {
	if (value === undefined) {
		return undefined;
	}
	if (value === null) {
		return null;
	}
	return {
		tenantId: value.tenantId,
		externalKey: value.externalKey,
		apiKey: value.apiKey,
		apiSecret: value.apiSecret,
		auditLogs:
			value.auditLogs === undefined
				? undefined
				: (value.auditLogs as Array<any>).map(AuditLogToJSON)
	};
}